Fixed slidingWindows function and warp bounding box
I was not able to validate that the `slidingWindows()` function worked
previously because I couldn't undistort images, so now I was able to see
the current x position of the windows was not named correctly and see
the results were as expected.

I changed the coordinates of the warp `src_corners` to make sure the
lines were fully within the image and parallel. I also removed the edge
detection because a sufficient binary image was created using just color
thresholding.