Improved README and added necessary files
The standard README that came with the project contained a bunch of 
information that was only relevant to someone completing the project. I 
removed that and added a little extra information to improve the quality 
of the README. I also began tracking files that are useful for running 
the project and would be needed by any cloning the repo.